build update for new files

+++ build.py
@@ -10,12 +10,12 @@ path_cur = os.getcwd();
 path_dist = "_dist/";
 path_dist_pkg = path_dist + "whatshow_phy_mod_otfs/";
 # file
-file = "OTFS.py"
+files = ["OTFS.py", "OTFSResGrid.py", "OTFSDetector.py"];
 file_init = "__init__.py";
 file_setup = "setup.py";
 file_readme = "README.md";
 # file content
-file_init_content = "from .OTFS import OTFS";
+file_init_content = "from .OTFS import OTFS\nfrom .OTFSResGrid import OTFSResGrid\nfrom .OTFSDetector import OTFSDetector";
 file_setup_content = f'from setuptools import setup, find_packages\n\
 \n\
 # prepare the instruction\n\
@@ -40,7 +40,9 @@ if os.path.exists(path_dist):
 os.makedirs(path_dist_pkg);   
 
 # copy files to the distribution
-shutil.copyfile(file, path_dist_pkg + file);
+for file in files:
+    if os.path.exists(file):
+        shutil.copyfile(file, path_dist_pkg + file);
 
 # add config files
 # __init__.py
@@ -56,7 +58,8 @@ with open(file_readme, "r") as readme:
     description = readme.read(); 
     txt_1_start = description.find("## How to install");
     txt_2_start = description.find("## How to use");
-    description_pypi = description[:txt_1_start] + description[txt_2_start:];
+    txt_3_start = description.find("## Samples");
+    description_pypi = description[:txt_1_start] + description[txt_2_start:txt_3_start];
 with open(path_dist + file_readme, "w") as readme_pypi:
     readme_pypi.write(description_pypi);
